Moved `plot_learning_curves` method out of the train script
